requirements.txt
# Core packages
aiohttp
annotated-types
contextily
cubo
datahugger
dask
dbt-core>=1.9.0
dbt-duckdb>=1.9.0
dbt-ibis>=0.11.0
duckdb>=1.3.1
duckdb-engine
earthengine-api
# esda
fiona
folium
fsspec
gdal
geoarrow-rs  # Rust implementation with better performance and I/O support
geoarrow-rust-core<= 0.4.0b3 # 0.4.0 release removes from/to geopandas and other utils
geoarrow-rust-compute
geoarrow-rust-io
geopandas
geodatasets
geopy
geos
graphviz
sf-hamilton
# optional but recommended: 
sf-hamilton[cli]
h3-py>4.1.0
ibis-framework[duckdb]
ibis-framework[geospatial]
ipyleaflet
ipython
ipywidgets
# jupyter_bbox_widget
jupysql>=0.11
jupyterlab
lonboard
matplotlib
maxar-platform
motherduck
notebook
numpy
obstore
# opencv-python
# openeo
overturemaps
overturemaestro  # Efficient Overture Maps data access
pandas
pillow
planetary-computer
pooch
# pvlib
pycountry
pydeck
# pygeohash
pyiceberg>=0.9.1
pyproj
pystac
pystac-client
python-dotenv
pytorch-lightning
pyyaml
rasterio
# rastervision
rich
rioxarray
rtree
sciencebasepy
scikit-image
scikit-learn
# scipy
seaborn
segmentation-models-pytorch~=0.5
shapely
stac-geoparquet
stackstac
tifffile
tqdm
torch>=2.0.0
torchgeo~=0.7
torchmetrics>=1.0.0
torchvision
urllib3
# wandb
xarray